requirements_test.txt
# Test dependencies for ac_modbus Home Assistant integration
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-homeassistant-custom-component>=0.12.0

# Development tools
ruff>=0.1.0
black>=23.0.0
isort>=5.12.0

# Runtime dependencies
pymodbus>=3.6.0
